Updated NavView UI
Card widgets in the grid-view weren't displaying the background images properly. I altered the frame sizing of the image/parent views & removed the overlay text, text was placed below the card. Newly imported images where used to populate most of the card widgets and background was updated to display the RBG flag. Also linked the SecondNavView to the last card widget.